fix: fix UI for the details drawer
